添加 本地消息序号，存储发送图片本地命名

+++ IMiPhone/CategoryTableViewController.m
@@ -105,7 +105,7 @@
     else if ([cellId isEqualToString:@"focus"]) {
         personNum = [FriendDataProxy sharedProxy].focusTotal;
     }
-    cell.detailTextLabel.text = [NSString stringWithFormat:@"%i%@",personNum ,NSLocalizedString(@"Unit.Person", nil)];
+    cell.detailTextLabel.text = [NSString stringWithFormat:@"%li%@",(long)personNum ,NSLocalizedString(@"Unit.Person", nil)];
 
   return cell;
 }
